+++ src/test/resources/fixtures/processing/blade/class-directive-interpolation-noop/expected.blade.php
@@ -1,0 +1,4 @@
+<div @class([
+  'text-center {{ $activeClass }} p-4 flex bg-red-500 font-bold' => $active,
+  "text-center $activeClass p-4 flex bg-red-500 font-bold" => $active,
+])></div>