+++ src/test/resources/fixtures/processing/blade/class-directive/input.blade.php
@@ -1,0 +1,5 @@
+<div @class([
+  'text-center p-4 flex bg-red-500 font-bold' => $active,
+  'font-bold flex p-4',
+  $compact => 'tracking-wide text-xs px-2 py-0.5',
+])></div>