+++ src/test/resources/fixtures/processing/blade/component-attributes/expected.blade.php
@@ -1,0 +1,6 @@
+<x-button
+  class="flex bg-red-500 p-4 text-center font-bold"
+  :class="$active ? 'flex p-4 font-bold' : ''"
+>
+  Save
+</x-button>