+++ src/test/resources/fixtures/processing/blade/component-attributes/input.blade.php
@@ -1,0 +1,6 @@
+<x-button
+  class="text-center p-4 flex bg-red-500 font-bold"
+  :class="$active ? 'font-bold flex p-4' : ''"
+>
+  Save
+</x-button>