+++ src/test/resources/fixtures/processing/js/tagged-template/expected.js
@@ -1,0 +1,1 @@
+const button = tw`flex bg-red-500 p-4 text-center font-bold`