+++ src/test/resources/fixtures/processing/js/tagged-template/input.js
@@ -1,0 +1,1 @@
+const button = tw`text-center p-4 flex bg-red-500 font-bold`